requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
asgiref==3.7.2
    # via django
boto3==1.28.58
    # via -r requirements.in
botocore==1.31.58
    # via
    #   boto3
    #   s3transfer
build==1.0.3
    # via pip-tools
click==8.1.7
    # via pip-tools
cssselect2==0.7.0
    # via svglib
dj-database-url==2.1.0
    # via -r requirements.in
django==4.2.5
    # via
    #   -r requirements.in
    #   dj-database-url
    #   django-appconf
    #   django-classy-tags
    #   django-cms
    #   django-entangled
    #   django-filer
    #   django-formtools
    #   django-parler
    #   django-polymorphic
    #   django-sekizai
    #   django-select2
    #   django-storage-url
    #   django-storages
    #   django-taggit
    #   django-taggit-templatetags
    #   django-treebeard
    #   djangocms-admin-style
    #   djangocms-alias
    #   djangocms-frontend
    #   djangocms-versioning
    #   easy-thumbnails
django-appconf==1.0.5
    # via django-select2
django-classy-tags==4.1.0
    # via
    #   django-cms
    #   django-sekizai
django-cms==4.1.0rc4
    # via
    #   -r requirements.in
    #   djangocms-alias
    #   djangocms-attributes-field
    #   djangocms-blog
    #   djangocms-frontend
    #   djangocms-text-ckeditor
    #   djangocms-versioning
django-entangled==0.5.4
    # via djangocms-frontend
django-filer==3.1.0
    # via
    #   -r requirements.in
    #   djangocms-blog
    #   djangocms-frontend
django-formtools==2.4.1
    # via django-cms
django-fsm==2.8.1
    # via djangocms-versioning
django-meta==2.4.0
    # via djangocms-blog
django-parler==2.3
    # via
    #   djangocms-alias
    #   djangocms-blog
django-polymorphic==3.1.0
    # via django-filer
django-sekizai==4.1.0
    # via django-cms
django-select2==8.1.2
    # via djangocms-frontend
django-sortedm2m==3.1.1
    # via djangocms-blog
django-storage-url==0.10.1
    # via -r requirements.in
django-storages==1.14.1
    # via django-storage-url
django-taggit==4.0.0
    # via
    #   django-taggit-autosuggest
    #   django-taggit-templatetags
    #   djangocms-blog
django-taggit-autosuggest==0.4.1
    # via djangocms-blog
django-taggit-templatetags==0.2.5
    # via djangocms-blog
django-templatetag-sugar==1.0
    # via django-taggit-templatetags
django-treebeard==4.7
    # via django-cms
djangocms-admin-style==3.2.6
    # via
    #   -r requirements.in
    #   django-cms
djangocms-alias==2.0.0rc2
    # via -r requirements.in
djangocms-attributes-field==3.0.0
    # via djangocms-frontend
djangocms-blog @ git+https://github.com/fsbraun/djangocms-blog@develop-4
    # via -r requirements.in
djangocms-frontend==1.1.8
    # via -r requirements.in
djangocms-text-ckeditor==5.1.3
    # via
    #   -r requirements.in
    #   djangocms-blog
    #   djangocms-frontend
djangocms-versioning==2.0.0rc1
    # via -r requirements.in
easy-thumbnails[svg]==2.8.5
    # via
    #   -r requirements.in
    #   django-filer
    #   djangocms-blog
    #   djangocms-frontend
furl==2.1.3
    # via django-storage-url
html5lib==1.1
    # via djangocms-text-ckeditor
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
lxml==4.9.3
    # via
    #   djangocms-blog
    #   svglib
orderedmultidict==1.0.1
    # via furl
packaging==23.2
    # via
    #   build
    #   django-cms
    #   djangocms-text-ckeditor
pillow==10.0.1
    # via
    #   djangocms-text-ckeditor
    #   easy-thumbnails
    #   reportlab
pip-tools==7.3.0
    # via -r requirements.in
psycopg2==2.9.8
    # via -r requirements.in
pyproject-hooks==1.0.0
    # via build
python-dateutil==2.8.2
    # via botocore
pytz==2023.3.post1
    # via djangocms-blog
reportlab==4.0.5
    # via
    #   easy-thumbnails
    #   svglib
s3transfer==0.7.0
    # via boto3
six==1.16.0
    # via
    #   furl
    #   html5lib
    #   orderedmultidict
    #   python-dateutil
sqlparse==0.4.4
    # via django
svglib==1.5.1
    # via easy-thumbnails
tinycss2==1.2.1
    # via
    #   cssselect2
    #   svglib
typing-extensions==4.8.0
    # via dj-database-url
urllib3==1.26.17
    # via botocore
uwsgi==2.0.22
    # via -r requirements.in
webencodings==0.5.1
    # via
    #   cssselect2
    #   html5lib
    #   tinycss2
wheel==0.41.2
    # via pip-tools
whitenoise==6.5.0
    # via -r requirements.in

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
